refactor: address code review — verbose module, generateAgentId to worker.ts, singleton init
Three related architectural improvements:

- verbose: moved out of StatusBar into its own src/agent/verbose.ts module
  (a config value that never changes doesn't belong as instance state). Both
  display.ts and status-bar.ts now import it from there.

- generateAgentId / WORKER_NAMES: moved from status-bar.ts to worker.ts where
  the worker identity concept belongs. To avoid a circular dependency (worker.ts
  already imports from status-bar.ts), the StatusBar constructor now requires
  agentId to be passed explicitly.

- statusBar singleton: status-bar.ts initialises with a plain UUID; index.ts
  calls initStatusBar(new StatusBar({ agentId: generateAgentId() })) early at
  startup so the live ES-module binding updates before any display calls run.
  Tests that don't call initStatusBar() work fine with the default instance.
