+++ CLAUDE.md
@@ -19,7 +19,7 @@ MVC structure. Models own state; controllers handle external inputs.
 
 A unified REPL + worker loop. `main()` in `index.ts` runs both interactive and worker modes; `startWorkerMode()` in `worker.ts` sets up the WebSocket session and returns before `main()` takes over the query loop.
 
-Key files: `worker.ts` (WS protocol + task lifecycle; also owns `generateAgentId` and `WORKER_NAMES`), `display.ts` (TUI rendering: colors, print helpers, diff/markdown rendering), `status-bar.ts` (status bar state and rendering; exports `StatusBar` class — holds all worker status state — and a shared `statusBar` singleton initialized via `initStatusBar()` at startup; callers import directly from this module, not via `display.ts`), `verbose.ts` (module-level `verbose` flag + `setVerbose()` — set once at startup from config, imported by `display.ts` and `status-bar.ts`), `command-registry.ts` (`CommandRegistry` class — instantiated in `index.ts` and injected into all modules; supports scoped sub-registries via `scoped(prefix)`; `registry.listCommandNames()` and `registry.listCommands()` return all available commands including file-based commands and skills), `workspace.ts` (git/npm workspace management), `worker-prompts.ts` (prompt templates).
+Key files: `worker.ts` (WS protocol + task lifecycle; also owns `generateAgentId` and `WORKER_NAMES`), `display.ts` (TUI rendering: colors, print helpers, diff/markdown rendering; also owns the `verbose` flag + `setVerbose()` alongside `thinkOutLoud`), `status-bar.ts` (status bar state and rendering; exports `StatusBar` class — holds all worker status state — and a shared `statusBar` singleton initialized via `initStatusBar()` at startup; callers import directly from this module, not via `display.ts`), `command-registry.ts` (`CommandRegistry` class — instantiated in `index.ts` and injected into all modules; supports scoped sub-registries via `scoped(prefix)`; `registry.listCommandNames()` and `registry.listCommands()` return all available commands including file-based commands and skills), `workspace.ts` (git/npm workspace management), `worker-prompts.ts` (prompt templates).
 
 ### Shared
