+++ src/agent/index.ts
@@ -6,9 +6,8 @@ import { fileURLToPath } from "url";
 import { query } from "@anthropic-ai/claude-agent-sdk";
 import type { CanUseTool, PermissionMode, PermissionResult } from "@anthropic-ai/claude-agent-sdk";
 import * as display from "./display.js";
-import { setThinkOutLoud } from "./display.js";
+import { setThinkOutLoud, setVerbose, verbose } from "./display.js";
 import { StatusBar, statusBar, initStatusBar } from "./status-bar.js";
-import { setVerbose, verbose } from "./verbose.js";
 import { ask, pick, pickMultiple, pickQuestion } from "./input.js";
 import type { PickQuestionResult } from "./input.js";
 import { WorkerSession, registerWorkerCommands, startWorkerMode, generateAgentId } from "./worker.js";
