+++ src/agent/status-bar.ts
@@ -1,6 +1,6 @@
 import { EventEmitter } from "node:events";
 import { shortWorkerId } from "../../shared/utils.js";
-import { verbose } from "./verbose.js";
+import { verbose } from "./display.js";
 import type { EffortValue } from "./effort.js";
 
 // ── Worker status types ────────────────────────────────────────────────────────
