+++ src/agent/worker.ts
@@ -6,7 +6,7 @@ import os from "node:os";
 import { WebSocket } from "ws";
 import * as display from "./display.js";
 import { StatusBar, statusBar } from "./status-bar.js";
-import { setVerbose } from "./verbose.js";
+import { setVerbose } from "./display.js";
 import { buildInitialPrompt, buildEventPrompt } from "./worker-prompts.js";
 import type { EffortValue } from "./effort.js";
 import * as Wire from "../../shared/wire.js";
