+++ tests/repl.foreman-message.test.ts
@@ -3,9 +3,8 @@
  * This ensures no message arrives silently, even when the worker is busy running a query.
  */
 import { describe, it, expect, vi, beforeEach, afterEach, afterAll } from "vitest";
-import { setVerbose, verbose } from "../src/agent/verbose.js";
+import { setVerbose, verbose, printForemanMessage } from "../src/agent/display.js";
 import { stripAnsi } from "./helpers.js";
-import { printForemanMessage } from "../src/agent/display.js";
 import { statusBar } from "../src/agent/status-bar.js";
 import * as Wire from "../shared/wire.js";
