+++ tests/repl.printing.test.ts
@@ -1,5 +1,4 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { setVerbose, verbose } from "../src/agent/verbose.js";
 import { stripAnsi } from "./helpers.js";
 import {
   printBlock,
@@ -7,6 +6,7 @@ import {
   print,
   toolUseNames,
   setThinkOutLoud,
+  setVerbose, verbose,
 } from "../src/agent/display.js";
 import { statusBar } from "../src/agent/status-bar.js";
